Move muiTheme generation to a utility function
